requirements.txt
django==5.2.1
markdown==3.8
python-dotenv==1.1.0
kagglehub==0.3.12
